+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/TableRowSelection.jsx
@@ -109,7 +109,7 @@ const TableRowSelection = () => {
         rowHeaderIndex={rowHeaderIndex}
         columnWidth="180px"
         ariaLabel="Table with Row Selections"
-        hasSelectableRows={hasSelectableRows} // Prop to turn row selection mode on/off
+        rowSelectionMode={hasSelectableRows ? 'multiple' : undefined} // Prop to turn row selection mode on/off
         onRowSelect={onRowSelect} // For row selection, consumer must provide a callback that the Worklist Data Grid will call when the user selects one or more rows.
         onRowSelectionHeaderSelect={onRowSelectionHeaderSelect}
       />