+++ packages/terra-table/src/subcomponents/Section.jsx
@@ -56,10 +56,10 @@ const propTypes = {
   rows: PropTypes.arrayOf(rowShape),
 
   /**
-   * Boolean indicating whether or not the table allows a row to be selected. If true, an additional
-   * column containing a checkbox is rendered to indicate when when the row is selected.
+   * Enables row selection capabilities for the table.
+   * Use 'single' for single row selection and 'multiple' for multi-row selection.
    */
-  hasRowSelection: PropTypes.bool,
+  rowSelectionMode: PropTypes.string,
 
   /**
    * All columns currently displayed.
@@ -111,7 +111,7 @@ function Section(props) {
     isHidden,
     isTableStriped,
     text,
-    hasRowSelection,
+    rowSelectionMode,
     displayedColumns,
     onCellSelect,
     rowHeight,
@@ -142,14 +142,11 @@ function Section(props) {
           role="columnheader"
           scope="col"
         >
-          <div>
-            <SectionHeader
-              text={text}
-              isOpen={!isCollapsed}
-              onClick={isCollapsible && onSectionSelect ? handleMouseDown : undefined}
-            />
-          </div>
-          <div />
+          <SectionHeader
+            text={text}
+            isOpen={!isCollapsed}
+            onClick={isCollapsible && onSectionSelect ? handleMouseDown : undefined}
+          />
         </th>
       </tr>
       )}
@@ -169,7 +166,7 @@ function Section(props) {
             tableId={tableId}
             cells={row.cells}
             ariaLabel={row.ariaLabel}
-            hasRowSelection={hasRowSelection}
+            rowSelectionMode={rowSelectionMode}
             displayedColumns={displayedColumns}
             rowHeaderIndex={rowHeaderIndex}
             onCellSelect={onCellSelect}