+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 * Added
- * Added examples and tests for `terra-folder-tree`.
+  * Added examples and tests to `terra-table` for single row selection mode.
+  * Added examples and tests for `terra-folder-tree`.
 
 ## 1.45.0 - (November 20, 2023)